+++ core/src/main/scala/org/broadinstitute/dsde/rawls/billing/BillingProjectLifecycle.scala
@@ -1,7 +1,9 @@
 package org.broadinstitute.dsde.rawls.billing
 
+import bio.terra.profile.model.ProfileModel
 import com.typesafe.scalalogging.LazyLogging
 import org.broadinstitute.dsde.rawls.RawlsExceptionWithErrorReport
+import org.broadinstitute.dsde.rawls.billing.BillingProfileManagerDAO.ProfilePolicy
 import org.broadinstitute.dsde.rawls.config.MultiCloudWorkspaceConfig
 import org.broadinstitute.dsde.rawls.dataaccess.SamDAO
 import org.broadinstitute.dsde.rawls.dataaccess.slick.WorkspaceManagerResourceMonitorRecord.JobType.JobType
@@ -15,7 +17,7 @@ import org.broadinstitute.dsde.rawls.model.{
 }
 
 import java.util.UUID
-import scala.concurrent.{ExecutionContext, Future}
+import scala.concurrent.{blocking, ExecutionContext, Future}
 
 /**
  * Handles provisioning and deleting billing projects with external providers. Implementors of this trait are not concerned
@@ -25,9 +27,10 @@ import scala.concurrent.{ExecutionContext, Future}
  */
 trait BillingProjectLifecycle extends LazyLogging {
 
-  // It's probably reasonable to expect that the billing project lifecyle always includes these two resources
+  // Resources common to all implementations.
   val samDAO: SamDAO
   val billingRepository: BillingRepository
+  val billingProfileManagerDAO: BillingProfileManagerDAO
 
   // The type of WorkspaceManagerResourceMonitorRecord job that should be created to finalize deletion when necessary
   val deleteJobType: JobType
@@ -74,6 +77,40 @@ trait BillingProjectLifecycle extends LazyLogging {
     executionContext: ExecutionContext
   ): Future[Unit]
 
+  def createBillingProfile(createProjectRequest: CreateRawlsV2BillingProjectFullRequest, ctx: RawlsRequestContext)(
+    implicit executionContext: ExecutionContext
+  ): Future[ProfileModel] =
+    Future(blocking {
+      val policies: Map[String, List[(String, String)]] =
+        if (createProjectRequest.protectedData.getOrElse(false)) Map("protected-data" -> List[(String, String)]())
+        else Map.empty
+      val profileModel = billingProfileManagerDAO.createBillingProfile(
+        createProjectRequest.projectName.value,
+        createProjectRequest.billingInfo,
+        policies,
+        ctx
+      )
+      logger.info(
+        s"Creating BPM-backed billing project ${createProjectRequest.projectName.value}, created profile with ID ${profileModel.getId}."
+      )
+      profileModel
+    })(executionContext)
+
+  def addMembersToBillingProfile(profileModel: ProfileModel,
+                                 createProjectRequest: CreateRawlsV2BillingProjectFullRequest,
+                                 ctx: RawlsRequestContext
+  )(implicit executionContext: ExecutionContext): Future[Set[Unit]] = {
+    val members = createProjectRequest.members.getOrElse(Set.empty)
+    Future.traverse(members) { member =>
+      Future(blocking {
+        billingProfileManagerDAO.addProfilePolicyMember(profileModel.getId,
+                                                        ProfilePolicy.fromProjectRole(member.role),
+                                                        member.email,
+                                                        ctx
+        )
+      })
+    }
+  }
 }
 
 class DuplicateBillingProjectException(errorReport: ErrorReport) extends RawlsExceptionWithErrorReport(errorReport)
