+++ core/src/main/scala/org/broadinstitute/dsde/rawls/billing/GoogleBillingProjectLifecycle.scala
@@ -1,8 +1,6 @@
 package org.broadinstitute.dsde.rawls.billing
 
 import akka.http.scaladsl.model.StatusCodes
-import bio.terra.profile.model.ProfileModel
-import org.broadinstitute.dsde.rawls.billing.BillingProfileManagerDAO.ProfilePolicy
 import org.broadinstitute.dsde.rawls.config.MultiCloudWorkspaceConfig
 import org.broadinstitute.dsde.rawls.dataaccess.slick.WorkspaceManagerResourceMonitorRecord.JobType.{
   GoogleBillingProjectDelete,
@@ -25,7 +23,7 @@ import org.broadinstitute.dsde.rawls.user.UserService.{
 }
 
 import java.util.UUID
-import scala.concurrent.{blocking, ExecutionContext, Future}
+import scala.concurrent.{ExecutionContext, Future}
 
 class GoogleBillingProjectLifecycle(
   val billingRepository: BillingRepository,
@@ -63,47 +61,13 @@ class GoogleBillingProjectLifecycle(
   override def postCreationSteps(createProjectRequest: CreateRawlsV2BillingProjectFullRequest,
                                  config: MultiCloudWorkspaceConfig,
                                  ctx: RawlsRequestContext
-  ): Future[CreationStatus] = {
-    val projectName = createProjectRequest.projectName
-
-    // This is where we create a billing profile in the BPM case.
-    def createBillingProfile: Future[ProfileModel] =
-      Future(blocking {
-        val policies: Map[String, List[(String, String)]] =
-          if (createProjectRequest.protectedData.getOrElse(false)) Map("protected-data" -> List[(String, String)]())
-          else Map.empty
-        val profileModel = billingProfileManagerDAO.createBillingProfile(
-          projectName.value,
-          createProjectRequest.billingInfo,
-          policies,
-          ctx
-        )
-        logger.info(
-          s"Creating BPM-backed billing project ${projectName.value}, created profile with ID ${profileModel.getId}."
-        )
-        profileModel
-      })
-
-    def addMembersToBillingProfile(profileModel: ProfileModel): Future[Set[Unit]] = {
-      val members = createProjectRequest.members.getOrElse(Set.empty)
-      Future.traverse(members) { member =>
-        Future(blocking {
-          billingProfileManagerDAO.addProfilePolicyMember(profileModel.getId,
-                                                          ProfilePolicy.fromProjectRole(member.role),
-                                                          member.email,
-                                                          ctx
-          )
-        })
-      }
-    }
-
+  ): Future[CreationStatus] =
     for {
-      - <- createBillingProfile.flatMap { profileModel =>
-        addMembersToBillingProfile(profileModel)
+      - <- createBillingProfile(createProjectRequest, ctx).flatMap { profileModel =>
+        addMembersToBillingProfile(profileModel, createProjectRequest, ctx)
       }
       _ <- syncBillingProjectOwnerPolicyToGoogleAndGetEmail(samDAO, createProjectRequest.projectName)
     } yield CreationStatuses.Ready
-  }
 
   override def initiateDelete(projectName: RawlsBillingProjectName, ctx: RawlsRequestContext)(implicit
     executionContext: ExecutionContext
